+++ src/methods/index.ts
@@ -1,6 +1,9 @@
 import { confirmPaymentIntentByCard } from "./confirmPaymentIntentByCard";
 import { addSourceToCustomer } from "./addSourceToCustomer";
 import { deleteSourceFromCustomer } from "./deleteSourceFromCustomer";
+import { getAllCards } from "./getAllCards";
+import { getCustomer } from "./getCustomer";
+import { setDefaultCard } from "./setDefaultCard";
 
 /**
  * additional stripe methods
@@ -13,4 +16,7 @@ export class RemedyProductStripe {
   public confirmPaymentIntentByCard = confirmPaymentIntentByCard;
   public addSourceToCustomer = addSourceToCustomer;
   public deleteSourceFromCustomer = deleteSourceFromCustomer;
+  public getAllCards = getAllCards;
+  public getCustomer = getCustomer;
+  public setDefaultCard = setDefaultCard;
 }